libsensors: Treat devices without a known ancestor bus as virtual
Currently only hwmon devices that don't have a parent device are treated
as virtual. Let's extend the concept to hwmon devices, that don't have
a *recognized ancestor* device (hwmon devices that according to the kernel
don't reside on a bus that we recognize). This change is meant to address
cases where a hwmon device has a thermal class device for a parent, but
the thermal class device doesn't have a parent device (these kind of
hwmon devices started appearing in the 4.19 kernel due to commit
f6b6b52ef7a54160c0).

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>